+++ src/App.jsx
@@ -3,6 +3,7 @@ import Home from './pages/Home'
 import DeckDetail from './pages/DeckDetail'
 import Review from './pages/Review'
 import Browse from './pages/Browse'
+import PromptGuide from './pages/PromptGuide'
 
 export default function App() {
   return (
@@ -12,6 +13,7 @@ export default function App() {
         <Route path="/deck/:id" element={<DeckDetail />} />
         <Route path="/review/:id" element={<Review />} />
         <Route path="/browse/:id" element={<Browse />} />
+        <Route path="/prompt-guide" element={<PromptGuide />} />
       </Routes>
     </HashRouter>
   )